fixed an issue uploading a file from data upload modal
