+++ client/src/components/forms/select/autocomplete/component.tsx
@@ -109,9 +109,10 @@ const AutoCompleteSelect = <T,>({
         >
           {({ selected, disabled }) => (
             <div
-              className={classnames('block truncate text-gray-900', {
+              className={classnames('block truncate', {
                 'text-navy-400': selected,
                 'text-gray-300': disabled,
+                'text-gray-900': !selected && !disabled,
               })}
             >
               {filteredOptions[index].label}
